move some codes to data service
